+++ src/main/java/dnd/myOcean/controller/letter/LetterController.java
@@ -1,0 +1,37 @@
+package dnd.myOcean.controller.letter;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/letter")
+public class LetterController {
+
+//    private final LetterService letterService;
+//
+//    /**
+//     * TODO : 편지 전송 기능
+//     */
+//    @AssignCurrentEmail
+//    @PostMapping
+//    public ResponseEntity sendLetter(LetterSendRequest request) {
+//        letterService.sendLetter(request);
+//
+//    }
+//
+//
+//    /**
+//     * TODO : 편지 삭제 기능
+//     */
+//
+//    /**
+//     * TODO : 편지 조회 기능
+//     */
+//    @AssignCurrentEmail
+//    @GetMapping
+//
+//    @AssignCurrentEmail
+//    @GetMapping("/send")
+}
