+++ src/main/java/dnd/myOcean/domain/letter/Letter.java
@@ -33,7 +33,7 @@ public class Letter extends BaseEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @OnDelete(action = OnDeleteAction.CASCADE)
-    @JoinColumn(name = "receiver_id")
+    @JoinColumn(name = "sender_id")
     private Member sender;
 
     @ManyToOne(fetch = FetchType.LAZY)
