+++ src/main/java/dnd/myOcean/dto/letter/request/LetterSendRequest.java
@@ -1,5 +1,6 @@
-package dnd.myOcean.dto.member;
+package dnd.myOcean.dto.letter.request;
 
+import dnd.myOcean.domain.member.Gender;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Null;
 import lombok.AccessLevel;
@@ -10,11 +11,17 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberNicknameUpdateRequest {
+public class LetterSendRequest {
 
     @Null
     private String email;
 
     @NotEmpty
-    private String nickname;
+    private String content;
+
+    private Gender gender;
+
+    private Integer ageRangeStart;
+
+    private Integer ageRangeEnd;
 }
