+++ src/main/java/dnd/myOcean/dto/letter/response/LetterSendResponse.java
@@ -1,0 +1,4 @@
+package dnd.myOcean.dto.letter.response;
+
+public class LetterSendResponse {
+}
