+++ src/main/java/dnd/myOcean/dto/member/request/MemberBirthdayUpdateRequest.java
@@ -1,16 +1,22 @@
-package dnd.myOcean.dto.member;
+package dnd.myOcean.dto.member.request;
 
+
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Null;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberInfoRequest {
+public class MemberBirthdayUpdateRequest {
 
     @Null
     private String email;
+
+    @NotEmpty
+    private String birthday;
 }
