+++ src/main/java/dnd/myOcean/dto/member/request/MemberGenderUpdateRequest.java
@@ -1,0 +1,21 @@
+package dnd.myOcean.dto.member.request;
+
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Null;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MemberGenderUpdateRequest {
+
+    @Null
+    private String email;
+
+    @NotEmpty
+    private String gender;
+}
