+++ src/main/java/dnd/myOcean/dto/member/request/MemberInfoRequest.java
@@ -1,22 +1,16 @@
-package dnd.myOcean.dto.member;
+package dnd.myOcean.dto.member.request;
 
-
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Null;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberBirthdayUpdateRequest {
+public class MemberInfoRequest {
 
     @Null
     private String email;
-
-    @NotEmpty
-    private String birthday;
 }
