+++ src/main/java/dnd/myOcean/service/letter/LetterService.java
@@ -1,0 +1,31 @@
+package dnd.myOcean.service.letter;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LetterService {
+//
+//    private final MemberRepository memberRepository;
+//    private final LetterRepository letterRepository;
+//
+//    public LetterSendResponse sendLetter(LetterSendRequest request) {
+//        List<Member> members = filterReceiver(request);
+//
+//        Letter letter = Letter.builder()
+//                .sender()
+//                .content(request.getContent())
+//                .build();
+//
+//    }
+//
+//    private List<Member> filterReceiver(LetterSendRequest request) {
+//        return null;
+//    }
+//
+//    private int generateRandomReceiverCount() {
+//        return new Random().nextInt(5) + 1;
+//    }
+}
