+++ src/main/java/dnd/myOcean/service/member/MemberService.java
@@ -4,11 +4,11 @@ package dnd.myOcean.service.member;
 import dnd.myOcean.domain.member.Gender;
 import dnd.myOcean.domain.member.Member;
 import dnd.myOcean.domain.member.Worry;
-import dnd.myOcean.dto.member.MemberBirthdayUpdateRequest;
-import dnd.myOcean.dto.member.MemberGenderUpdateRequest;
-import dnd.myOcean.dto.member.MemberInfoRequest;
-import dnd.myOcean.dto.member.MemberNicknameUpdateRequest;
-import dnd.myOcean.dto.member.MemberWorryUpdateRequest;
+import dnd.myOcean.dto.member.request.MemberBirthdayUpdateRequest;
+import dnd.myOcean.dto.member.request.MemberGenderUpdateRequest;
+import dnd.myOcean.dto.member.request.MemberInfoRequest;
+import dnd.myOcean.dto.member.request.MemberNicknameUpdateRequest;
+import dnd.myOcean.dto.member.request.MemberWorryUpdateRequest;
 import dnd.myOcean.dto.member.response.MemberInfoResponse;
 import dnd.myOcean.exception.member.AlreadyExistNicknameException;
 import dnd.myOcean.exception.member.BirthdayUpdateLimitExceedException;
